+++ mount-vm-root.sh
@@ -389,7 +389,7 @@ dryrun_summary() {
 
 setup() {
     define_colours
-    PROJECT_VERSION="3.4.2"; SCRIPT_VERSION="3.0.0"
+    PROJECT_VERSION="3.4.3"; SCRIPT_VERSION="3.0.0"
     ROOT=""; MODE="--ro"
     parse_arguments "$@"
     check_elevation
@@ -859,7 +859,7 @@ dryrun_summary() {
 
 setup() {
     define_colours
-    PROJECT_VERSION="3.4.2"; SCRIPT_VERSION="3.0.0"
+    PROJECT_VERSION="3.4.3"; SCRIPT_VERSION="3.0.0"
     MOUNT_ROOT=""; MODE=""
     parse_arguments "$@"
     check_elevation
@@ -1300,7 +1300,7 @@ dryrun_summary() {
 
 setup() {
     define_colours
-    PROJECT_VERSION="3.4.2"; SCRIPT_VERSION="3.0.0"
+    PROJECT_VERSION="3.4.3"; SCRIPT_VERSION="3.0.0"
     MOUNT_ROOT=""; MODE="--ro"; KPARTX_STATUS=0
     parse_arguments "$@"
     check_elevation